test/behave/requirements.txt
behave==1.2.6
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
parse==1.19.0
parse-type==0.6.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.11
